Fix dependency injection in case RestAPI component is disabled
